MPTI test version

See merge request s08-webmobile1-sub2/S08P12A803!102

+++ backend/business/src/main/java/mpti/domain/opinion/api/response/GetReviewResponse.java
@@ -22,7 +22,7 @@ public class GetReviewResponse {
 
     private LocalDateTime createdAt;
 
-    private Double star;
+    private int star;
 
     public GetReviewResponse(Review review) {
         this.id = review.getId();
